Add files via upload
In the discipline of software engineering, a sequence diagram shows object interactions in temporal order. It represents the scenario's objects as well as the sequence of messages exchanged between them in order to carry out the scenario's functionality. In the logical view of the system under development, sequence diagrams are usually coupled with use case realizations. Event diagrams and event scenarios are other names for sequence diagrams.